get rid of legacy paynote
