+++ apps/www/components/Article/graphql/getDocument.js
@@ -215,8 +215,6 @@ export const getDocument = gql`
         indicateGallery
         indicateVideo
         prepublication
-        paynotes
-        paynoteMode
       }
     }
   }
