+++ packages/styleguide/src/components/SeriesNav/SeriesNavTile.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import { css } from 'glamor'
 
-import { InvertedColorScheme } from '../Colors/ColorContext'
 import {
   PADDING,
   TILE_MARGIN_RIGHT,
@@ -38,13 +37,6 @@ const styles = {
     minWidth: 150,
     maxWidth: 170,
   }),
-  paynoteLabelSpace: css({
-    marginTop: 0,
-    ...sansSerifRegular12,
-    [mUp]: {
-      ...sansSerifRegular14,
-    },
-  }),
 }
 
 const SeriesNavTile = ({
@@ -54,53 +46,29 @@ const SeriesNavTile = ({
   inline,
   ActionBar,
   Link,
-  index,
-  repoId,
-  context,
-  PayNote,
   onEpisodeClick,
   aboveTheFold,
 }) => {
-  const isInverted = !!PayNote
-  const content = PayNote ? (
-    <>
-      <p {...styles.paynoteLabelSpace}>&nbsp;</p>
-      <PayNote context={context} repoId={repoId} index={index} />
-    </>
-  ) : (
-    <SeriesNavTileContent
-      inline={inline}
-      episode={episode}
-      current={current}
-      ActionBar={ActionBar}
-      Link={Link}
-      onEpisodeClick={onEpisodeClick}
-      aboveTheFold={aboveTheFold}
-      t={t}
-    />
-  )
-  const inactiveTile = !isInverted && !episode?.document?.meta?.path
+  const inactiveTile = !episode?.document?.meta?.path
 
   return (
     <div
-      /* 
-      Hide PayNote for members by setting the data-hide-if-active-membership attribute. 
-      This class is defined in the MeContext in republik-frontend
-      */
-      {...(PayNote && {
-        'data-hide-if-active-membership': true,
-      })}
       {...(inline ? styles.inlineTile : styles.tile)}
       style={{
         opacity: inactiveTile ? 0.6 : 1,
       }}
       data-repo-id={episode?.document?.repoId}
     >
-      {isInverted ? (
-        <InvertedColorScheme>{content}</InvertedColorScheme>
-      ) : (
-        content
-      )}
+      <SeriesNavTileContent
+        inline={inline}
+        episode={episode}
+        current={current}
+        ActionBar={ActionBar}
+        Link={Link}
+        onEpisodeClick={onEpisodeClick}
+        aboveTheFold={aboveTheFold}
+        t={t}
+      />
     </div>
   )
 }
