+++ packages/styleguide/src/templates/Page/index.js
@@ -6,7 +6,6 @@ const createPageSchema = ({
   customMetaFields = [],
   series = false,
   darkMode,
-  paynotes = true,
   Link = DefaultLink,
   getPath = ({ slug }) => `/${(slug || '').split('/').pop()}`,
   metaHeadlines = true,
@@ -38,7 +37,6 @@ const createPageSchema = ({
     ],
     series,
     darkMode,
-    paynotes,
     Link,
     metaHeadlines,
     skipContainer,
